+++ src/jpeg_encoder.hpp
@@ -5,5 +5,5 @@
 
 class JpegEncoder {
 public:
-	static std::vector<uint8_t> encodeRgba(const uint8_t *rgba, int width, int height, int quality);
+	static std::vector<uint8_t> encodeBgr(const uint8_t *bgr, int width, int height, int quality);
 };